app/requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
pdf2image
pymupdf
python-multipart